refactor(http): transfer adapters to a package
Sprint 9 Review 2
